// src/lib/enums/product.enum.ts
export enum ProductStatus{
    PAUSE = "PAUSE",
    PROCESS = "PROCESS",
    DELETE = "DELETE"
}

export enum ProductSize{
    SMALL = "SMALL",
    NORMAL = "NORMAL",
    LARGE = "LARGE",
    SET = "SET",

}

export enum ProductCollection{
    DISH = "DISH",
    SALAD = "SALAD",
    DESERT = "DESERT",
    DRINK = "DRINK",
    OTHER = "OTHER"
}

export enum ProductVolume{
    ONE = 1,
    HALF = 0.5,
    ONE_POINT_TWO = 1.2,
    ONE_POINT_FIVE = 1.5,
    TWO = 2
}